Rename namespace Actions\Interfaces to Actions\Contracts
